- Bed flips are now clicked 3 times as it might be purchasable before it technically should be. For this a new config property "BED_MULTIPLE_CLICKS_DELAY" was added. It clicks once the configured milliseconds before it should be purchasable. Once at the exact time and once afterwards. It can be disabled to only click once by setting the property to a negative number. - fix chat messages not working if they contain apostrophes - added logic to claim auctions if they are expired - fix powershell script crashing if path contains a space - removed window skipping completely, because you get banned using it - added logic to enable the websocket to run sequences
Important:
I removed the .exe for windows users as it seems for some reason it was very unstable for some users. Windows users should now only use the .cmd or powershell script. These run BAF directly using Node.js.
Because of this the powershell script was rewritten, making the old version not work anymore.